change git-initializer e2e tests to integration using golang's testing pkg
